+++ DDHidLib/DDHidKeyboard.m
@@ -86,7 +86,7 @@
     return mKeyElements;
 }
 
-- (unsigned) numberOfKeys;
+- (NSUInteger) numberOfKeys;
 {
     return [mKeyElements count];
 }
