+++ DDHidLib/DDHidKeyboardBarcodeScanner.h
@@ -48,7 +48,7 @@
 
 - (NSArray *) keyElements;
 
-- (unsigned) numberOfKeys;
+- (NSUInteger) numberOfKeys;
 
 - (void) addElementsToQueue: (DDHidQueue *) queue;
